nvme/059: test startup with inaccessible namespaces
Inaccessible paths will not be visible to the OS until at least
one path to the namespace becomes active (ie either 'optimized'
or 'non-optimized'). As there are reports that the system might
not be able to correctly register the namespace when all paths
started off as 'inaccessible' add a test for it.

Signed-off-by: Hannes Reinecke <hare@suse.de>